Refactoring: moved options to GPCloudOptions.kt
